+++ src/components/ServiceWorker.svelte
@@ -1,5 +1,7 @@
 <script>
   import { getContext, onMount } from 'svelte';
+
+  import { dev } from '$app/environment';
   import { page } from '$app/stores';
 
   import UpdateModal from './UpdateModal.svelte';
@@ -38,14 +40,18 @@
         if (event.data.type === 'update') refreshUpdate(event.data.changelog);
       });
 
-      navigator.serviceWorker.register('/service-worker.js').then(
-        function () {
-          console.log('service worker registration succeeded');
-        },
-        function (error) {
-          console.log('service worker registration failed:', error);
-        },
-      );
+      navigator.serviceWorker
+        .register('/service-worker.js', {
+          type: dev ? 'module' : 'classic',
+        })
+        .then(
+          function () {
+            console.log('service worker registration succeeded');
+          },
+          function (error) {
+            console.log('service worker registration failed:', error);
+          },
+        );
     } else {
       console.log('service workers are not supported');
     }
