+++ src/routes/+error.svelte
@@ -1,50 +1,38 @@
-<script context="module">
-  export function load({ error, status }) {
-    return {
-      props: {
-        status,
-        error,
-      },
-    };
-  }
-</script>
-
 <script>
-  import { onMount } from 'svelte';
+  import { page } from '$app/stores';
 
-  export let status;
-  export let error;
+  import { onMount } from 'svelte';
 
   const dev = import.meta.env.DEV;
 
   let refreshUrl;
 
   onMount(() => {
-    console.error('ERROR', error);
+    console.error('ERROR', $page.error);
     refreshUrl = `${location.href}?r=${new Date().getTime()}`;
   });
 </script>
 
 <svelte:head>
-  <title>{status}</title>
+  <title>{$page.status}</title>
 </svelte:head>
 
 <div class="lg:ml-64 p-16 flex flex-col justify-center items-center">
   <div class="relative">
-    <h1 class="bg-red-500 text-red-700 rounded-lg px-4 font-display font-black" style="font-size: 6rem;">{status}</h1>
+    <h1 class="bg-red-500 text-red-700 rounded-lg px-4 font-display font-black" style="font-size: 6rem;">{$page.status}</h1>
     <h1
       class="absolute top-0 left-0 text-background-secondary font-display font-black text-6xl"
       style="margin-left: 14px; z-index: 0; font-size: 6rem;"
     >
-      {status}
+      {$page.status}
     </h1>
   </div>
 
   <p class="text-red-400 text-xl mt-4 text-center">
-    {status === 404 ? 'Page not found! Try one on the menu :)' : error.message}
+    {$page.status === 404 ? 'Page not found! Try one on the menu :)' : $page.error.message}
   </p>
 
-  {#if status === 500}
+  {#if $page.status === 500}
     <a
       class="px-4 py-2 rounded-xl border-2 border-white border-opacity-25 hover:border-primary text-blue-400 hover:text-primary text-xl mt-4 text-center"
       href={refreshUrl}
@@ -59,7 +47,7 @@
     </p>
   {/if}
 
-  {#if dev && error.stack}
-    <pre class="mt-8 p-4 rounded-lg text-black bg-red-400">{error.stack}</pre>
+  {#if dev && $page.error.stack}
+    <pre class="mt-8 p-4 rounded-lg text-black bg-red-400">{$page.error.stack}</pre>
   {/if}
 </div>
