+++ src/routes/+layout.js
@@ -1,0 +1,1 @@
+export const prerender = true;
