+++ src/routes/domains/[id]/+page.js
@@ -1,0 +1,9 @@
+import { domains } from '../../../data/domain.js';
+
+export async function load({ params }) {
+  const { id } = params;
+
+  const domain = domains[id];
+
+  return { id, domain };
+}
