+++ src/routes/weapons/+page.js
@@ -1,0 +1,1 @@
+import data from '../../data/weapons/en.json';
