+++ src/routes/wish/[id]/+page.js
@@ -1,0 +1,4 @@
+export async function load({ params }) {
+  const { id } = params;
+  return { id };
+}
