+++ src/routes/wish/import/+page.svelte
@@ -18,24 +18,24 @@
   import isSameOrBefore from 'dayjs/plugin/isSameOrBefore';
   dayjs.extend(isSameOrBefore);
 
-  import Button from '../../components/Button.svelte';
-  import Input from '../../components/Input.svelte';
-  import Select from '../../components/Select.svelte';
-  import Icon from '../../components/Icon.svelte';
-  import { getTimeOffset, server } from '../../stores/server';
-  import ImportFaqModal from './_importFaq.svelte';
-  import BackupReminderModal from './_backupReminder.svelte';
-  import Textarea from '../../components/Textarea.svelte';
-  import { getAccountPrefix } from '../../stores/account';
-  import { getLocalSaveJson, readSave, updateSave } from '../../stores/saveManager';
-  import { weaponList } from '../../data/weaponList';
-  import { characters } from '../../data/characters';
-  import { pushToast } from '../../stores/toast';
-  import { submitWishTally } from '../../functions/wishTally';
-  import Ad from '../../components/Ad.svelte';
-  import Checkbox from '../../components/Checkbox.svelte';
-  import { driveSignedIn } from '../../stores/dataSync';
-  import { banners } from '../../data/banners';
+  import Button from '../../../components/Button.svelte';
+  import Input from '../../../components/Input.svelte';
+  import Select from '../../../components/Select.svelte';
+  import Icon from '../../../components/Icon.svelte';
+  import { getTimeOffset, server } from '../../../stores/server';
+  import ImportFaqModal from '../_importFaq.svelte';
+  import BackupReminderModal from '../_backupReminder.svelte';
+  import Textarea from '../../../components/Textarea.svelte';
+  import { getAccountPrefix } from '../../../stores/account';
+  import { getLocalSaveJson, readSave, updateSave } from '../../../stores/saveManager';
+  import { weaponList } from '../../../data/weaponList';
+  import { characters } from '../../../data/characters';
+  import { pushToast } from '../../../stores/toast';
+  import { submitWishTally } from '../../../functions/wishTally';
+  import Ad from '../../../components/Ad.svelte';
+  import Checkbox from '../../../components/Checkbox.svelte';
+  import { driveSignedIn } from '../../../stores/dataSync';
+  import { banners } from '../../../data/banners';
 
   const numberFormat = Intl.NumberFormat();
   let fetchController;
